ajustes

+++ Common/DTOs/DtoHistorialCambios.cs
@@ -14,8 +14,8 @@ namespace Common.DTOs
         public DateTime fechaIngreso { get; set; }
         public string observaciones { get; set; }
         public string comentario { get; set; }
-        public decimal latitud { get; set; }
-        public decimal longitud { get; set; }
+        public double latitud { get; set; }
+        public double longitud { get; set; }
         public string estado { get; set; }
         public DtoReclamo dtoReclamo { get; set; }
